Explicity set the group leader if specified in params
Since the 2017-03-11 release changing the leader of a Space has been broken. This appears to be due to https://github.com/instructure/canvas-lms/commit/a9fab4f4af26c468c6196351ccca7ef499730aa5 -- when calling params.slice(*SETTABLE_GROUP_ATTRIBUTES), the `leader` attribute is not present. We now explicitly set the leader of the group if a :leader param is present.